+++ binance/client.py
@@ -124,7 +124,7 @@ class BaseClient:
 
     def __init__(
         self, api_key: Optional[str] = None, api_secret: Optional[str] = None,
-        requests_params: Dict[str, str] = None, tld: str = 'com',
+        requests_params: Optional[Dict[str, str]] = None, tld: str = 'com',
         testnet: bool = False
     ):
         """Binance API Client constructor
@@ -217,7 +217,7 @@ class BaseClient:
         return url + '/' + self.OPTIONS_API_VERSION + '/' + path
 
     def _generate_signature(self, data: Dict) -> str:
-
+        assert self.API_SECRET, "API Secret required for private endpoints"
         ordered_data = self._order_params(data)
         query_string = '&'.join([f"{d[0]}={d[1]}" for d in ordered_data])
         m = hmac.new(self.API_SECRET.encode('utf-8'), query_string.encode('utf-8'), hashlib.sha256)
@@ -290,7 +290,7 @@ class Client(BaseClient):
 
     def __init__(
         self, api_key: Optional[str] = None, api_secret: Optional[str] = None,
-        requests_params: Dict[str, str] = None, tld: str = 'com',
+        requests_params: Optional[Dict[str, str]] = None, tld: str = 'com',
         testnet: bool = False
     ):
 
@@ -883,6 +883,8 @@ class Client(BaseClient):
             return self.get_klines(**params)
         elif HistoricalKlinesType.FUTURES == klines_type:
             return self.futures_klines(**params)
+        elif HistoricalKlinesType.FUTURES_COIN == klines_type:
+            return self.futures_coin_klines(**params)
         else:
             raise NotImplementedException(klines_type)
 
@@ -909,7 +911,7 @@ class Client(BaseClient):
         )
         return kline[0][0]
 
-    def get_historical_klines(self, symbol, interval, start_str, end_str=None, limit=500,
+    def get_historical_klines(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                               klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
         """Get Historical Klines from Binance
 
@@ -917,21 +919,23 @@ class Client(BaseClient):
         :type symbol: str
         :param interval: Binance Kline interval
         :type interval: str
-        :param start_str: Start date string in UTC format or timestamp in milliseconds
+        :param start_str: optional - start date string in UTC format or timestamp in milliseconds
         :type start_str: str|int
         :param end_str: optional - end date string in UTC format or timestamp in milliseconds (default will fetch everything up to now)
         :type end_str: str|int
-        :param limit: Default 500; max 1000.
+        :param limit: Default 1000; max 1000.
         :type limit: int
         :param klines_type: Historical klines type: SPOT or FUTURES
         :type klines_type: HistoricalKlinesType
 
-        :return: list of OHLCV values
+        :return: list of OHLCV values (Open time, Open, High, Low, Close, Volume, Close time, Quote asset volume, Number of trades, Taker buy base asset volume, Taker buy quote asset volume, Ignore)
 
         """
-        return self._historical_klines(symbol, interval, start_str, end_str=end_str, limit=limit, klines_type=klines_type)
+        return self._historical_klines(
+            symbol, interval, start_str=start_str, end_str=end_str, limit=limit, klines_type=klines_type
+        )
 
-    def _historical_klines(self, symbol, interval, start_str, end_str=None, limit=500,
+    def _historical_klines(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                            klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
         """Get Historical Klines from Binance (spot or futures)
 
@@ -943,16 +947,16 @@ class Client(BaseClient):
         :type symbol: str
         :param interval: Binance Kline interval
         :type interval: str
-        :param start_str: Start date string in UTC format or timestamp in milliseconds
+        :param start_str: optional - start date string in UTC format or timestamp in milliseconds
         :type start_str: str|int
         :param end_str: optional - end date string in UTC format or timestamp in milliseconds (default will fetch everything up to now)
         :type end_str: None|str|int
-        :param limit: Default 500; max 1000.
+        :param limit: Default 1000; max 1000.
         :type limit: int
         :param klines_type: Historical klines type: SPOT or FUTURES
         :type klines_type: HistoricalKlinesType
 
-        :return: list of OHLCV values
+        :return: list of OHLCV values (Open time, Open, High, Low, Close, Volume, Close time, Quote asset volume, Number of trades, Taker buy base asset volume, Taker buy quote asset volume, Ignore)
 
         """
         # init our list
@@ -961,14 +965,18 @@ class Client(BaseClient):
         # convert interval to useful value in seconds
         timeframe = interval_to_milliseconds(interval)
 
+        # if a start time was passed convert it
         start_ts = convert_ts_str(start_str)
 
         # establish first available start timestamp
-        first_valid_ts = self._get_earliest_valid_timestamp(symbol, interval, klines_type)
-        start_ts = max(start_ts, first_valid_ts)
+        if start_ts is not None:
+            first_valid_ts = self._get_earliest_valid_timestamp(symbol, interval, klines_type)
+            start_ts = max(start_ts, first_valid_ts)
 
         # if an end time was passed convert it
         end_ts = convert_ts_str(end_str)
+        if end_ts and start_ts and end_ts <= start_ts:
+            return output_data
 
         idx = 0
         while True:
@@ -982,32 +990,31 @@ class Client(BaseClient):
                 endTime=end_ts
             )
 
-            # handle the case where exactly the limit amount of data was returned last loop
-            if not len(temp_data):
-                break
-
             # append this loops data to our output data
-            output_data += temp_data
+            if temp_data:
+                output_data += temp_data
 
-            # set our start timestamp using the last value in the array
-            start_ts = temp_data[-1][0]
-
-            idx += 1
+            # handle the case where exactly the limit amount of data was returned last loop
             # check if we received less than the required limit and exit the loop
-            if len(temp_data) < limit:
+            if not len(temp_data) or len(temp_data) < limit:
                 # exit the while loop
                 break
 
             # increment next call by our timeframe
-            start_ts += timeframe
+            start_ts = temp_data[-1][0] + timeframe
+
+            # exit loop if we reached end_ts before reaching <limit> klines
+            if end_ts and start_ts >= end_ts:
+                break
 
             # sleep after every 3rd call to be kind to the API
+            idx += 1
             if idx % 3 == 0:
                 time.sleep(1)
 
         return output_data
 
-    def get_historical_klines_generator(self, symbol, interval, start_str, end_str=None,
+    def get_historical_klines_generator(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                                         klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
         """Get Historical Klines generator from Binance
 
@@ -1015,10 +1022,12 @@ class Client(BaseClient):
         :type symbol: str
         :param interval: Binance Kline interval
         :type interval: str
-        :param start_str: Start date string in UTC format or timestamp in milliseconds
+        :param start_str: optional - Start date string in UTC format or timestamp in milliseconds
         :type start_str: str|int
         :param end_str: optional - end date string in UTC format or timestamp in milliseconds (default will fetch everything up to now)
         :type end_str: str|int
+        :param limit: amount of candles to return per request (default 1000)
+        :type limit: int
         :param klines_type: Historical klines type: SPOT or FUTURES
         :type klines_type: HistoricalKlinesType
 
@@ -1026,9 +1035,9 @@ class Client(BaseClient):
 
         """
 
-        return self._historical_klines_generator(symbol, interval, start_str, end_str=end_str, klines_type=klines_type)
+        return self._historical_klines_generator(symbol, interval, start_str, end_str, limit, klines_type=klines_type)
 
-    def _historical_klines_generator(self, symbol, interval, start_str, end_str=None,
+    def _historical_klines_generator(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                                      klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
         """Get Historical Klines generator from Binance (spot or futures)
 
@@ -1040,7 +1049,7 @@ class Client(BaseClient):
         :type symbol: str
         :param interval: Binance Kline interval
         :type interval: str
-        :param start_str: Start date string in UTC format or timestamp in milliseconds
+        :param start_str: optional - Start date string in UTC format or timestamp in milliseconds
         :type start_str: str|int
         :param end_str: optional - end date string in UTC format or timestamp in milliseconds (default will fetch everything up to now)
         :type end_str: str|int
@@ -1050,21 +1059,22 @@ class Client(BaseClient):
         :return: generator of OHLCV values
 
         """
-        # setup the max limit
-        limit = 500
 
         # convert interval to useful value in seconds
         timeframe = interval_to_milliseconds(interval)
 
-        # convert our date strings to milliseconds
+        # if a start time was passed convert it
         start_ts = convert_ts_str(start_str)
 
         # establish first available start timestamp
-        first_valid_ts = self._get_earliest_valid_timestamp(symbol, interval, klines_type)
-        start_ts = max(start_ts, first_valid_ts)
+        if start_ts is not None:
+            first_valid_ts = self._get_earliest_valid_timestamp(symbol, interval, klines_type)
+            start_ts = max(start_ts, first_valid_ts)
 
         # if an end time was passed convert it
         end_ts = convert_ts_str(end_str)
+        if end_ts and start_ts and end_ts <= start_ts:
+            return
 
         idx = 0
         while True:
@@ -1078,27 +1088,27 @@ class Client(BaseClient):
                 endTime=end_ts
             )
 
-            # handle the case where exactly the limit amount of data was returned last loop
-            if not len(output_data):
-                break
-
             # yield data
-            for o in output_data:
-                yield o
+            if output_data:
+                for o in output_data:
+                    yield o
 
-            # set our start timestamp using the last value in the array
-            start_ts = output_data[-1][0]
-
-            idx += 1
+            # handle the case where exactly the limit amount of data was returned last loop
             # check if we received less than the required limit and exit the loop
-            if len(output_data) < limit:
+            if not len(output_data) or len(output_data) < limit:
                 # exit the while loop
                 break
 
+            # set our start timestamp using the last value in the array
             # increment next call by our timeframe
-            start_ts += timeframe
+            start_ts = output_data[-1][0] + timeframe
+
+            # exit loop if we reached end_ts before reaching <limit> klines
+            if end_ts and start_ts >= end_ts:
+                break
 
             # sleep after every 3rd call to be kind to the API
+            idx += 1
             if idx % 3 == 0:
                 time.sleep(1)
 
@@ -2163,6 +2173,35 @@ class Client(BaseClient):
 
         """
         return self._request_margin_api('get', 'account/apiRestrictions', True, data=params)
+
+    def get_dust_assets(self, **params):
+        """Get assets that can be converted into BNB
+
+        https://binance-docs.github.io/apidocs/spot/en/#get-assets-that-can-be-converted-into-bnb-user_data
+
+        :returns: API response
+
+        .. code-block:: python
+
+            {
+                "details": [
+                    {
+                        "asset": "ADA",
+                        "assetFullName": "ADA",
+                        "amountFree": "6.21",   //Convertible amount
+                        "toBTC": "0.00016848",  //BTC amount
+                        "toBNB": "0.01777302",  //BNB amount（Not deducted commission fee）
+                        "toBNBOffExchange": "0.01741756", //BNB amount（Deducted commission fee）
+                        "exchange": "0.00035546" //Commission fee
+                    }
+                ],
+                "totalTransferBtc": "0.00016848",
+                "totalTransferBNB": "0.01777302",
+                "dribbletPercentage": "0.02"     //Commission fee
+            }
+
+        """
+        return self._request_margin_api('post', 'asset/dust-btc', True, data=params)
 
     def get_dust_log(self, **params):
         """Get log of small amounts exchanged for BNB.
@@ -3429,6 +3468,67 @@ class Client(BaseClient):
         params['transTo'] = "ISOLATED_MARGIN"
         return self._request_margin_api('post', 'margin/isolated/transfer', signed=True, data=params)
 
+    def get_isolated_margin_tranfer_history(self, **params):
+        """Get transfers to isolated margin account.
+
+        https://binance-docs.github.io/apidocs/spot/en/#get-isolated-margin-transfer-history-user_data
+
+        :param asset: name of the asset
+        :type asset: str
+        :param symbol: pair required
+        :type symbol: str
+        :param transFrom: optional SPOT, ISOLATED_MARGIN
+        :param transFrom: str SPOT, ISOLATED_MARGIN
+        :param transTo: optional
+        :param transTo: str
+        :param startTime: optional
+        :type startTime: int
+        :param endTime: optional
+        :type endTime: int
+        :param current: Currently querying page. Start from 1. Default:1
+        :type current: str
+        :param size: Default:10 Max:100
+        :type size: int
+        :param recvWindow: the number of milliseconds the request is valid for
+        :type recvWindow: int
+
+        .. code:: python
+
+            transfer = client.transfer_spot_to_isolated_margin(symbol='ETHBTC')
+
+        :returns: API response
+
+        .. code-block:: python
+
+            {
+              "rows": [
+                {
+                  "amount": "0.10000000",
+                  "asset": "BNB",
+                  "status": "CONFIRMED",
+                  "timestamp": 1566898617000,
+                  "txId": 5240372201,
+                  "transFrom": "SPOT",
+                  "transTo": "ISOLATED_MARGIN"
+                },
+                {
+                  "amount": "5.00000000",
+                  "asset": "USDT",
+                  "status": "CONFIRMED",
+                  "timestamp": 1566888436123,
+                  "txId": 5239810406,
+                  "transFrom": "ISOLATED_MARGIN",
+                  "transTo": "SPOT"
+                }
+              ],
+              "total": 2
+            }
+
+        :raises: BinanceRequestException, BinanceAPIException
+
+        """
+        return self._request_margin_api('get', 'margin/isolated/transfer', signed=True, data=params)
+
     def create_margin_loan(self, **params):
         """Apply for a loan in cross-margin or isolated-margin account.
 
@@ -3760,6 +3860,37 @@ class Client(BaseClient):
 
         """
         return self._request_margin_api('get', 'margin/repay', signed=True, data=params)
+
+    def get_cross_margin_data(self, **params):
+        """Query Cross Margin Fee Data (USER_DATA)
+
+        https://binance-docs.github.io/apidocs/spot/en/#query-cross-margin-fee-data-user_data
+        :param vipLevel: User's current specific margin data will be returned if vipLevel is omitted
+        :type vipLevel: int
+        :param coin
+        :type coin: str
+        :param recvWindow: the number of milliseconds the request is valid for
+        :type recvWindow: int
+        :returns: API response (example):
+            [
+                {
+                    "vipLevel": 0,
+                    "coin": "BTC",
+                    "transferIn": true,
+                    "borrowable": true,
+                    "dailyInterest": "0.00026125",
+                    "yearlyInterest": "0.0953",
+                    "borrowLimit": "180",
+                    "marginablePairs": [
+                        "BNBBTC",
+                        "TRXBTC",
+                        "ETHBTC",
+                        "BTCUSDT"
+                    ]
+                }
+            ]
+        """
+        return self._request_margin_api('get', 'margin/crossMarginData', signed=True, data=params)
 
     def get_margin_interest_history(self, **params):
         """Get Interest History (USER_DATA)
@@ -5705,7 +5836,7 @@ class Client(BaseClient):
         :param limit: Default 500; max 1000.
         :type limit: int
 
-        :return: list of OHLCV values
+        :return: list of OHLCV values (Open time, Open, High, Low, Close, Volume, Close time, Quote asset volume, Number of trades, Taker buy base asset volume, Taker buy quote asset volume, Ignore)
 
         """
         return self._historical_klines(symbol, interval, start_str, end_str=end_str, limit=limit, klines_type=HistoricalKlinesType.FUTURES)
@@ -6010,7 +6141,7 @@ class Client(BaseClient):
 
         """
         params = {
-            'true' if multiAssetsMargin else 'false'
+            'multiAssetsMargin': 'true' if multiAssetsMargin else 'false'
         }
         return self._request_futures_api('post', 'multiAssetsMargin', True, data=params)
 
@@ -6020,7 +6151,7 @@ class Client(BaseClient):
         https://binance-docs.github.io/apidocs/futures/en/#get-current-multi-assets-mode-user_data
 
         """
-        return self._request_futures_api('get', 'multiAssetsMargin', True)
+        return self._request_futures_api('get', 'multiAssetsMargin', True, data={})
 
     def futures_stream_get_listen_key(self):
         res = self._request_futures_api('post', 'listenKey', signed=False, data={})
@@ -7107,7 +7238,7 @@ class AsyncClient(BaseClient):
 
     def __init__(
         self, api_key: Optional[str] = None, api_secret: Optional[str] = None,
-        requests_params: Dict[str, str] = None, tld: str = 'com',
+        requests_params: Optional[Dict[str, str]] = None, tld: str = 'com',
         testnet: bool = False, loop=None
     ):
 
@@ -7117,7 +7248,7 @@ class AsyncClient(BaseClient):
     @classmethod
     async def create(
         cls, api_key: Optional[str] = None, api_secret: Optional[str] = None,
-        requests_params: Dict[str, str] = None, tld: str = 'com',
+        requests_params: Optional[Dict[str, str]] = None, tld: str = 'com',
         testnet: bool = False, loop=None
     ):
 
@@ -7346,6 +7477,8 @@ class AsyncClient(BaseClient):
             return await self.get_klines(**params)
         elif HistoricalKlinesType.FUTURES == klines_type:
             return await self.futures_klines(**params)
+        elif HistoricalKlinesType.FUTURES_COIN == klines_type:
+            return await self.futures_coin_klines(**params)
         else:
             raise NotImplementedException(klines_type)
     _klines.__doc__ = Client._klines.__doc__
@@ -7363,12 +7496,12 @@ class AsyncClient(BaseClient):
         return kline[0][0]
     _get_earliest_valid_timestamp.__doc__ = Client._get_earliest_valid_timestamp.__doc__
 
-    async def get_historical_klines(self, symbol, interval, start_str, end_str=None, limit=500,
+    async def get_historical_klines(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                                     klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
         return await self._historical_klines(symbol, interval, start_str, end_str=end_str, limit=limit, klines_type=klines_type)
     get_historical_klines.__doc__ = Client.get_historical_klines.__doc__
 
-    async def _historical_klines(self, symbol, interval, start_str, end_str=None, limit=500,
+    async def _historical_klines(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                                  klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
 
         # init our list
@@ -7377,15 +7510,16 @@ class AsyncClient(BaseClient):
         # convert interval to useful value in seconds
         timeframe = interval_to_milliseconds(interval)
 
-        # convert our date strings to milliseconds
-        start_ts = convert_ts_str(start_str)
-
         # establish first available start timestamp
-        first_valid_ts = await self._get_earliest_valid_timestamp(symbol, interval, klines_type)
-        start_ts = max(start_ts, first_valid_ts)
+        start_ts = convert_ts_str(start_str)
+        if start_ts is not None:
+            first_valid_ts = await self._get_earliest_valid_timestamp(symbol, interval, klines_type)
+            start_ts = max(start_ts, first_valid_ts)
 
         # if an end time was passed convert it
         end_ts = convert_ts_str(end_str)
+        if end_ts and start_ts and end_ts <= start_ts:
+            return output_data
 
         idx = 0
         while True:
@@ -7399,55 +7533,57 @@ class AsyncClient(BaseClient):
                 endTime=end_ts
             )
 
-            # handle the case where exactly the limit amount of data was returned last loop
-            if not len(temp_data):
-                break
-
             # append this loops data to our output data
-            output_data += temp_data
+            if temp_data:
+                output_data += temp_data
 
-            # set our start timestamp using the last value in the array
-            start_ts = temp_data[-1][0]
-
-            idx += 1
-            # check if we received less than the required limit and exit the loop
-            if len(temp_data) < limit:
+            # handle the case where exactly the limit amount of data was returned last loop
+            # or check if we received less than the required limit and exit the loop
+            if not len(temp_data) or len(temp_data) < limit:
                 # exit the while loop
                 break
 
-            # increment next call by our timeframe
-            start_ts += timeframe
+            # set our start timestamp using the last value in the array
+            # and increment next call by our timeframe
+            start_ts = temp_data[-1][0] + timeframe
+
+            # exit loop if we reached end_ts before reaching <limit> klines
+            if end_ts and start_ts >= end_ts:
+                break
 
             # sleep after every 3rd call to be kind to the API
+            idx += 1
             if idx % 3 == 0:
                 await asyncio.sleep(1)
 
         return output_data
     _historical_klines.__doc__ = Client._historical_klines.__doc__
 
-    async def get_historical_klines_generator(self, symbol, interval, start_str, end_str=None,
+    async def get_historical_klines_generator(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                                               klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
-        return self._historical_klines_generator(symbol, interval, start_str, end_str=end_str, klines_type=klines_type)
+        return self._historical_klines_generator(
+            symbol, interval, start_str, end_str=end_str, limit=limit, klines_type=klines_type
+        )
     get_historical_klines_generator.__doc__ = Client.get_historical_klines_generator.__doc__
 
-    async def _historical_klines_generator(self, symbol, interval, start_str, end_str=None,
+    async def _historical_klines_generator(self, symbol, interval, start_str=None, end_str=None, limit=1000,
                                            klines_type: HistoricalKlinesType = HistoricalKlinesType.SPOT):
-
-        # setup the max limit
-        limit = 500
 
         # convert interval to useful value in seconds
         timeframe = interval_to_milliseconds(interval)
 
-        # convert our date strings to milliseconds
+        # if a start time was passed convert it
         start_ts = convert_ts_str(start_str)
 
         # establish first available start timestamp
-        first_valid_ts = await self._get_earliest_valid_timestamp(symbol, interval, klines_type)
-        start_ts = max(start_ts, first_valid_ts)
+        if start_ts is not None:
+            first_valid_ts = await self._get_earliest_valid_timestamp(symbol, interval, klines_type)
+            start_ts = max(start_ts, first_valid_ts)
 
         # if an end time was passed convert it
         end_ts = convert_ts_str(end_str)
+        if end_ts and start_ts and end_ts <= start_ts:
+            return
 
         idx = 0
         while True:
@@ -7461,27 +7597,26 @@ class AsyncClient(BaseClient):
                 endTime=end_ts
             )
 
-            # handle the case where exactly the limit amount of data was returned last loop
-            if not len(output_data):
-                break
-
             # yield data
-            for o in output_data:
-                yield o
+            if output_data:
+                for o in output_data:
+                    yield o
 
-            # set our start timestamp using the last value in the array
-            start_ts = output_data[-1][0]
-
-            idx += 1
+            # handle the case where exactly the limit amount of data was returned last loop
             # check if we received less than the required limit and exit the loop
-            if len(output_data) < limit:
+            if not len(output_data) or len(output_data) < limit:
                 # exit the while loop
                 break
 
             # increment next call by our timeframe
-            start_ts += timeframe
+            start_ts = output_data[-1][0] + timeframe
+
+            # exit loop if we reached end_ts before reaching <limit> klines
+            if end_ts and start_ts >= end_ts:
+                break
 
             # sleep after every 3rd call to be kind to the API
+            idx += 1
             if idx % 3 == 0:
                 await asyncio.sleep(1)
     _historical_klines_generator.__doc__ = Client._historical_klines_generator.__doc__
@@ -7623,6 +7758,10 @@ class AsyncClient(BaseClient):
     async def get_account_api_permissions(self, **params):
         return await self._request_margin_api('get', 'account/apiRestrictions', True, data=params)
     get_account_api_permissions.__doc__ = Client.get_account_api_permissions.__doc__
+
+    async def get_dust_assets(self, **params):
+        return await self._request_margin_api('post', 'asset/dust-btc', True, data=params)
+    get_dust_assets.__doc__ = Client.get_dust_assets.__doc__
 
     async def get_dust_log(self, **params):
         return await self._request_margin_api('get', 'asset/dribblet', True, data=params)
@@ -7786,6 +7925,9 @@ class AsyncClient(BaseClient):
 
     async def get_margin_repay_details(self, **params):
         return await self._request_margin_api('get', 'margin/repay', signed=True, data=params)
+
+    async def get_cross_margin_data(self, **params):
+        return await self._request_margin_api('get', 'margin/crossMarginData', signed=True, data=params)
 
     async def get_margin_interest_history(self, **params):
         return await self._request_margin_api('get', 'margin/interestHistory', signed=True, data=params)
@@ -8005,7 +8147,7 @@ class AsyncClient(BaseClient):
         return await self._request_futures_api('get', 'continuousKlines', data=params)
 
     async def futures_historical_klines(self, symbol, interval, start_str, end_str=None, limit=500):
-        return self._historical_klines(symbol, interval, start_str, end_str=end_str, limit=limit, klines_type=HistoricalKlinesType.FUTURES)
+        return await self._historical_klines(symbol, interval, start_str, end_str=end_str, limit=limit, klines_type=HistoricalKlinesType.FUTURES)
 
     async def futures_historical_klines_generator(self, symbol, interval, start_str, end_str=None):
         return self._historical_klines_generator(symbol, interval, start_str, end_str=end_str, klines_type=HistoricalKlinesType.FUTURES)
@@ -8117,12 +8259,12 @@ class AsyncClient(BaseClient):
 
     async def futures_change_multi_assets_mode(self, multiAssetsMargin: bool):
         params = {
-            'true' if multiAssetsMargin else 'false'
+            'multiAssetsMargin': 'true' if multiAssetsMargin else 'false'
         }
         return await self._request_futures_api('post', 'multiAssetsMargin', True, data=params)
 
     async def futures_get_multi_assets_mode(self):
-        return await self._request_futures_api('get', 'multiAssetsMargin', True)
+        return await self._request_futures_api('get', 'multiAssetsMargin', True, data={})
 
     async def futures_stream_get_listen_key(self):
         res = await self._request_futures_api('post', 'listenKey', signed=False, data={})