+++ binance/exceptions.py
@@ -82,3 +82,7 @@ class NotImplementedException(Exception):
     def __init__(self, value):
         message = f'Not implemented: {value}'
         super().__init__(message)
+
+
+class UnknownDateFormat(Exception):
+    ...